// src/src/types/request.d.ts
interface IAppSReportItem {
    objectId: number;
    objectType?: string | null;
    objectName?: string | null;
    isFavourite: boolean;
    pageNumber: number;
    pageSize: number;
    pagination: IPagination | any;
    documentationURL?: string;
    applicationURL?: string;
}

interface IPagination {
    totalRows: number;
    totalPages: number;
}